Mac (Unix/X11) でコンパイルできない問題を修正しました。

+++ src/autokey.c
@@ -25,6 +25,7 @@ based on https://github.com/windy6001/ROMAJI/
 */
 
 
+#include <string.h>
 #include "buffer.h"
 
 static RINGBUFFER* autokeybuffer;
@@ -55,7 +56,7 @@ int getAutokey(unsigned char *data)
 
 int putAutokeyMessage(unsigned char* data)
 {
-    for (int i = 0; i < strlen(data); i++) {
+    for (size_t i = 0; i < strlen(data); i++) {
         ringbuffer_Put(autokeybuffer, data[i]);
     }
